requirements.txt
Django>=5.1,<6.1
djangorestframework>=3.15,<4.0
django-filter>=24.0,<26.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
celery>=5.4,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0
python-dotenv>=1.0,<2.0
PyMySQL>=1.1,<2.0
certifi>=2024.8.30,<2027.0
